maybe

+++ src/mbr.rs
@@ -22,7 +22,7 @@ pub fn list_partitions(disk: String) {
 
 /// creates a basic partition table then formats the disk
 /// * if make_swap is set to true it creates a partition table with a swap of the specifed size.
-pub fn basic_arch_part(user_disk: String, make_swap: bool, swap_size: u32) {
+pub fn basic_arch_part(user_disk: String, make_swap: bool, swap_size: usize) {
     let mut f = std::fs::File::create(&user_disk).expect("could not open disk");
     let mut mbr  = mbrman::MBR::new_from(&mut f, 512, [0x01, 0x02, 0x03, 0x04])
 	.expect("could not make partition table");
@@ -58,7 +58,7 @@ pub fn basic_arch_part(user_disk: String, make_swap: bool, swap_size: u32) {
 	sys: 0x82,
 	last_chs: mbrman::CHS::empty(),
 	starting_lba,
-	sectors: swap_size,
+	sectors: swap_size as u32,
     };
 
     let free_partition_number = mbr.iter().find(|(i, p)| p.is_unused()).map(|(i, _)| i)
